+++ src/CreateMiddleware/CreateMiddleware.php
@@ -22,6 +22,8 @@ class CreateMiddleware
     public const CLASS_SKELETON = <<< 'EOS'
 <?php
 
+declare(strict_types=1);
+
 namespace %namespace%;
 
 use Psr\Http\Message\ResponseInterface;
@@ -39,6 +41,7 @@ class %class% implements MiddlewareInterface
         // $response = $handler->handle($request);
     }
 }
+
 EOS;
 
     /**