+++ src/Module/Create.php
@@ -9,12 +9,12 @@ declare(strict_types=1);
 
 namespace Zend\Expressive\Tooling\Module;
 
-use Zend\Expressive\Tooling\Module\Exception;
-
 class Create
 {
     public const TEMPLATE_CONFIG_PROVIDER = <<< 'EOT'
 <?php
+
+declare(strict_types=1);
 
 namespace %1$s;